Move the Nitro SDK task classes to a separate 'tasks' folder
